Recommender/Backend/API/api_requirements.txt
Flask
mysql-connector-python
pymongo
pandas